requirements.txt
SQLAlchemy~=1.4.29
sqlalchemy-utils~=0.38.2
concurrent-log-handler~=0.9.19
psycopg2-binary~=2.9.3
alembic~=1.7.5
coloredlogs~=15.0.1
PyYAML~=6.0
tornado~=6.1
pytz~=2021.3
requests~=2.27.1
PyJWT~=2.3.0
deeply~=1.0.3
